move HelloController to controller folder
