+++ src/main/resources/application.properties
@@ -1,0 +1,2 @@
+spring.datasource.url=jdbc:h2:mem:test;DB_CLOSE_ON_EXIT=FALSE
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.H2Dialect
