+++ src/test/java/com/example/springboot/controller/HelloControllerIT.java
@@ -1,4 +1,4 @@
-package com.example.springboot;
+package com.example.springboot.controller;
 
 import org.junit.jupiter.api.Test;
